pyproject.toml
[project]
name = "astro-virtual-assistant"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "quart>=0.20.0",
    "quart-schema[pydantic]>=0.21.0",
    "quart-injector==0.0.0b1",
    "pydantic>=2.10.5",
    "logstash_formatter>=0.5.17",
    "aiohttp>=3.11.11",
    "aioprometheus[aiohttp]>=23.12.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.4",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "aioresponses>=0.7.8",
    "pytest-mock-resources[docker]>=2.12.1",
    "pytest-env>=1.1.5",
    "hypothesis>=6.125.3",
    "coverage>=7.6.12",
    "openapi-spec-validator>=0.7.1",
    "syrupy>=4.9.1",
]

[tool.uv.workspace]
members = ["libs/*", "services/*"]
